requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
brotlipy==0.7.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
fastapi==0.89.1
fonttools==4.38.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
httpx
idna==3.4
isort==5.11.4
kiwisolver==1.4.4
matplotlib==3.6.2
multidict==6.0.4
munkres==1.1.4
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.1
packaging==23.0
pandas==1.5.2
pathspec==0.10.3
Pillow==9.4.0
pip==22.3.1
platformdirs==2.6.2
pycparser==2.21
pydantic==1.10.4
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytest
pytz==2022.7.1
PyYAML==6.0
redis
requests==2.28.1
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchsummary==1.5.1
torchvision==0.14.1
tornado==6.2
types-docutils==0.19.1.1
types-Pillow==9.4.0.3
types-setuptools==65.7.0.1
types-tabulate==0.9.0.0
types-tqdm==4.64.7.10
typing_extensions==4.4.0
unicodedata2==15.0.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wheel==0.38.4
yarl==1.8.2

# db setup
databases[postgresql]
SQLAlchemy
alembic
psycopg2

# combining celery tasks and async functions
asgiref